+++ src/VarDumperInterface.php
@@ -16,9 +16,18 @@ interface VarDumperInterface
     public function dump($var);
 
     /**
+     * @deprecated
+     *
      * @param mixed $var
      *
      * @return string
      */
     public function getDump($var);
+
+    /**
+     * @param mixed $var
+     *
+     * @return string
+     */
+    public function dumpAsString($var);
 }